refactor: extract lifecycle and shortcuts hooks
